Fetching data from server and put on to the view pager
View Pager
The ViewPager is the one widget of android that allows the user to swipe
left or right to see an entirely new screen.
In a sense, it's just a nicer way to show the user multiple tabs. It
also has the ability to dynamically add and remove pages (or tabs) at
anytime.
Consider the example of grouping search results by employee list
categories, and showing each employee details in a separate list. With
the ViewPager, the application user could then swipe left or right to
see other employee details
Using the ViewPager requires some knowledge of both Fragments and
PageAdapters.
In this case, Fragments are "pages".
Each screen that the ViewPager allows the user to scroll to is really a
Fragment. By using Fragments instead of a View here,
we're given a much wider range of possibilities to show in each page.
We're not limited to just a List of items. This could be any collection
of views and widgets we may need.
You can think of PageAdapters in the same way that you think of
ListAdapters. The main difference between page adapter and list adapter
is that the Page Adapter's job is to supply Fragments (instead of views)
to the UI for drawing.

Fragment
A fragment is a module of code that holds part of the behavior and/or UI
of an activity. As the name intimates, fragments are not independent
entities, but are sub parts to a single activity. In many ways, they
resemble and echo the functionality of activities.
Suppose the scenario of you have lots of details of user  and it is
quite difficult to understand when all such information are comes under
the single view,  to handle such kind of complexity fragment can be
used,`
Fragment can be used to categorized the employee details and display the
details into fragment according to the category .
Fragment  can provide:
Modularity: dividing complex activity code across fragments for better
organization and maintenance.
Reusability: placing behavior or UI parts into fragments that can be
shared across multiple activities.
Adaptability: representing sections of a UI as different fragments and
utilizing different layouts   depending on screen orientation and size.